refactor!: replace ProjectsProvider with RepositoySearchQuery
