--- src/main/kotlin/provider/ProjectsProvider.kt
@@ -1,14 +1,0 @@
-package provider
-
-import provider.repository.Repository
-
-/**
- * A provider of projects.
- */
-interface ProjectsProvider {
-
-    /**
-     * An [Iterable] of [Repository].
-     */
-    val repositories: Iterable<Repository>
-}
