+++ src/main/kotlin/provider/criteria/SearchCriteria.kt
@@ -1,0 +1,13 @@
+package provider.criteria
+
+/**
+ * An interface modeling a search criteria.
+ * @param T the result type returned applying the criteria.
+ */
+interface SearchCriteria<T> {
+    /**
+     * Apply the criteria.
+     * @return T the result of applying the criteria.
+     */
+    fun apply(): T
+}
