+++ src/main/kotlin/repository/AbstractRepository.kt
@@ -1,4 +1,4 @@
-package provider.repository
+package repository
 
 import org.apache.commons.io.FileUtils
 import org.apache.commons.io.filefilter.DirectoryFileFilter
