+++ src/test/kotlin/provider/ProviderTest.kt
@@ -5,12 +5,14 @@ import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.collections.shouldNotBeEmpty
 import io.kotest.matchers.collections.shouldNotContainDuplicates
-import io.kotest.matchers.ints.shouldBeExactly
+import io.kotest.matchers.nulls.shouldBeNull
+import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.string.shouldMatch
-import provider.query.ByBitbucketName
-import provider.query.ByBitbucketUser
-import provider.query.ByGitHubName
-import provider.query.ByGitHubUser
+import provider.criteria.ByBitbucketName
+import provider.criteria.ByBitbucketUser
+import provider.criteria.ByGitHubName
+import provider.criteria.ByGitHubUser
+import repository.Repository
 import java.net.URL
 
 private const val GH_URL_PREFIX = "https://github.com"
@@ -21,16 +23,19 @@ private const val BB_USER = "danysk"
 
 class ProviderTest : FunSpec() {
     init {
-        test("Searching by existing name and user should return repos matching that name") {
+        val githubProvider = GitHubProvider()
+        val bitbucketProvider = BitbucketProvider()
+
+        test("Searching by existing name and user should return repos matching those criteria") {
             var searchedRepoName = "Student-Project-OOP"
             testByExistingName(
-                GitHubProvider(ByGitHubName(searchedRepoName, ByGitHubUser(GH_USER))),
+                githubProvider.byCriteria(ByGitHubName(searchedRepoName, ByGitHubUser(GH_USER))),
                 searchedRepoName,
                 USERNAME
             )
             searchedRepoName = "OOP"
             testByExistingName(
-                BitbucketProvider(ByBitbucketName(searchedRepoName, ByBitbucketUser(BB_USER))),
+                bitbucketProvider.byCriteria(ByBitbucketName(searchedRepoName, ByBitbucketUser(BB_USER))),
                 searchedRepoName,
                 USERNAME
             )
@@ -39,57 +44,58 @@ class ProviderTest : FunSpec() {
         test("Searching by existing URL should return the repo pointed to") {
             var expectedRepoUrl = "$GH_URL_PREFIX/$GH_USER/code-plagiarism-detector"
             testByExistingUrl(
-                GitHubProvider(URL(expectedRepoUrl)),
+                githubProvider.byLink(URL(expectedRepoUrl)),
                 "code-plagiarism-detector",
                 USERNAME
             )
             expectedRepoUrl = "$BB_URL_PREFIX/$BB_USER/courses-oop-gradle-jfx-template"
             testByExistingUrl(
-                BitbucketProvider(URL(expectedRepoUrl)),
+                bitbucketProvider.byLink(URL(expectedRepoUrl)),
                 "Courses - OOP - Gradle JFX template",
                 USERNAME
             )
         }
 
-        test("Searching by *non-existing* URL should return an empty collection of repos") {
-            GitHubProvider(URL("$GH_URL_PREFIX/$GH_USER/non-existing-repo")).repositories.shouldBeEmpty()
-            BitbucketProvider(URL("$BB_URL_PREFIX/$BB_USER/non-existing-repo")).repositories.shouldBeEmpty()
+        test("Searching by *non-existing* URL should return null") {
+            githubProvider.byLink(URL("$GH_URL_PREFIX/$GH_USER/non-existing-repo")).shouldBeNull()
+            bitbucketProvider.byLink(URL("$BB_URL_PREFIX/$BB_USER/non-existing-repo")).shouldBeNull()
         }
 
         test("Searching by illegal URL should throw an exception") {
             shouldThrow<java.lang.IllegalArgumentException> {
-                GitHubProvider(URL("https://www.unibo.it/"))
+                githubProvider.byLink(URL("https://www.unibo.it/"))
             }
             shouldThrow<java.lang.IllegalArgumentException> {
-                GitHubProvider(URL("$GH_URL_PREFIX/$GH_USER/code-plagiarism-detector/tree/master/src/"))
+                githubProvider.byLink(URL("$GH_URL_PREFIX/$GH_USER/code-plagiarism-detector/tree/master/src/"))
             }
             shouldThrow<java.lang.IllegalArgumentException> {
-                BitbucketProvider(URL("https://www.unibo.it/"))
+                bitbucketProvider.byLink(URL("https://www.unibo.it/"))
             }
             shouldThrow<java.lang.IllegalArgumentException> {
-                BitbucketProvider(URL("$BB_URL_PREFIX/$BB_USER/courses-oop-gradle-jfx-template/src/master/"))
+                bitbucketProvider.byLink(URL("$BB_URL_PREFIX/$BB_USER/courses-oop-gradle-jfx-template/src/"))
             }
         }
 
         test("Searching by a *non-existing* name should return an empty collection of repos") {
-            GitHubProvider(ByGitHubName("non-existing-repo", ByGitHubUser(GH_USER)))
-                .repositories.shouldBeEmpty()
-            BitbucketProvider(ByBitbucketName("non-existing-repo", ByBitbucketUser(GH_USER)))
-                .repositories.shouldBeEmpty()
+            githubProvider.byCriteria(
+                ByGitHubName("non-existing-repo", ByGitHubUser(GH_USER))
+            ).shouldBeEmpty()
+            bitbucketProvider.byCriteria(
+                ByBitbucketName("non-existing-repo", ByBitbucketUser(GH_USER))
+            ).shouldBeEmpty()
         }
 
         test("Searching by a *non-existing* user should return an empty collection of repos") {
-            GitHubProvider(ByGitHubUser("non-existing-user")).repositories.shouldBeEmpty()
-            BitbucketProvider(ByBitbucketUser("non-existing-user")).repositories.shouldBeEmpty()
+            githubProvider.byCriteria(
+                ByGitHubUser("non-existing-user")
+            ).shouldBeEmpty()
+            bitbucketProvider.byCriteria(
+                ByBitbucketUser("non-existing-user")
+            ).shouldBeEmpty()
         }
     }
 
-    private fun testByExistingName(
-        provider: ProjectsProvider,
-        expectedName: String,
-        expectedUser: String
-    ) {
-        val result = provider.repositories
+    private fun testByExistingName(result: Iterable<Repository>, expectedName: String, expectedUser: String) {
         result.shouldNotBeEmpty()
         result.shouldNotContainDuplicates()
         result.forEach {
@@ -98,14 +104,9 @@ class ProviderTest : FunSpec() {
         }
     }
 
-    private fun testByExistingUrl(
-        provider: ProjectsProvider,
-        expectedName: String,
-        expectedUser: String
-    ) {
-        val result = provider.repositories
-        result.count().shouldBeExactly(1)
-        result.first().name.shouldMatch(expectedName)
-        result.first().owner.shouldMatch(expectedUser)
+    private fun testByExistingUrl(result: Repository?, expectedName: String, expectedUser: String) {
+        result.shouldNotBeNull()
+        result.name.shouldMatch(expectedName)
+        result.owner.shouldMatch(expectedUser)
     }
 }
