+++ src/test/kotlin/provider/RepositoryTest.kt
@@ -3,7 +3,7 @@ package provider
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.ints.shouldBeExactly
-import provider.repository.clonedReposDirectoryPath
+import repository.clonedReposDirectoryPath
 import java.io.File
 import java.net.URL
 
@@ -13,22 +13,24 @@ private const val EXPECTED_GH_REPO_SOURCES = 9
 private const val EXPECTED_BB_REPO_SOURCES = 3
 
 class RepositoryTest : FunSpec({
-    val githubProvider = GitHubProvider(URL(GH_SAMPLE_REPO_URL))
-    val bitbucketProvider = BitbucketProvider(URL(BB_SAMPLE_REPO_URL))
+    val githubRepo = GitHubProvider().byLink(URL(GH_SAMPLE_REPO_URL))!!
+    val bitbucketRepo = BitbucketProvider().byLink(URL(BB_SAMPLE_REPO_URL))!!
 
     test("Check if returns all sources") {
-        githubProvider.repositories.first().getSources("java").count() shouldBeExactly EXPECTED_GH_REPO_SOURCES
-        bitbucketProvider.repositories.first().getSources("java").count() shouldBeExactly EXPECTED_BB_REPO_SOURCES
+        githubRepo.getSources("java").count() shouldBeExactly EXPECTED_GH_REPO_SOURCES
+        bitbucketRepo.getSources("java").count() shouldBeExactly EXPECTED_BB_REPO_SOURCES
     }
 
     test("Trying to get sources of unknown language should throw an exception") {
         shouldThrow<java.lang.IllegalArgumentException> {
-            githubProvider.repositories.first().getSources("non-existing-language")
+            githubRepo.getSources("non-existing-language")
         }
         shouldThrow<java.lang.IllegalArgumentException> {
-            bitbucketProvider.repositories.first().getSources("non-existing-language")
+            bitbucketRepo.getSources("non-existing-language")
         }
     }
 
-    File(clonedReposDirectoryPath).deleteRecursively()
+    afterSpec {
+        File(clonedReposDirectoryPath).deleteRecursively()
+    }
 })
